UPDATE: move typed dispatch and selector hooks to hooks folder
