+++ src/components/EmployeeDetailsForm/EmployeeDetailsForm.tsx
@@ -1,6 +1,6 @@
 import { useState, useRef } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useAppSelector, useAppDispatch } from '../../core/store';
+import { useAppSelector, useAppDispatch } from '../../hooks/storeHelpers';
 import { Formik } from 'formik';
 import {
     Button,
