+++ src/components/EmployeesTableFilter/EmployeesTableFilter.tsx
@@ -1,4 +1,4 @@
-import { useAppDispatch, useAppSelector } from '../../core/store';
+import { useAppSelector, useAppDispatch } from '../../hooks/storeHelpers';
 import { Input, Button } from '..';
 import Select, { MultiValue } from 'react-select';
 import {
