+++ src/core/store/index.ts
@@ -8,7 +8,6 @@ import { composeWithDevTools } from '@redux-devtools/extension';
 import employeesReducer from './employeesList/reducer';
 import employeeReducer from './employee/reducer';
 import dropdownReducer from './dropdownData/reducer';
-import { useDispatch, useSelector, TypedUseSelectorHook } from 'react-redux';
 import { ThunkAction, thunk } from 'redux-thunk';
 
 const rootReducer = combineReducers({
@@ -26,10 +25,7 @@ const store = createStore(
 );
 
 export type RootState = ReturnType<typeof rootReducer>;
-
 export type AppDispatch = typeof store.dispatch;
-export const useAppDispatch: () => AppDispatch = useDispatch;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
 
 export type AppThunk<ReturnType = Promise<any>> = ThunkAction<
     ReturnType,
