+++ src/hooks/storeHelpers.ts
@@ -1,0 +1,5 @@
+import { useDispatch, TypedUseSelectorHook, useSelector } from 'react-redux';
+import { AppDispatch, RootState } from '../core/store';
+
+export const useAppDispatch: () => AppDispatch = useDispatch;
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
